kakarot_scripts/aws_lambda/relayers_observer/requirements.txt
aws-cdk-lib==2.161.1
constructs>=10.0.0,<11.0.0
